+++ vllm_rbln/platform.py
@@ -35,8 +35,8 @@ import vllm_rbln.rbln_envs as envs
 from vllm_rbln.logger import init_logger
 from vllm_rbln.utils.optimum.configuration import (
     is_qwen3_pooling,
-    sync_with_rbln_config,
 )
+from vllm_rbln.utils.optimum.converter import sync_vllm_and_optimum
 from vllm_rbln.utils.optimum.registry import (
     is_enc_dec_arch,
     is_multi_modal,
@@ -269,7 +269,7 @@ class RblnPlatform(Platform):
                     del model_config.__dict__["is_encoder_decoder"]
 
             cls.disable_unsupported_prefix_caching(vllm_config)
-            sync_with_rbln_config(vllm_config)
+            sync_vllm_and_optimum(vllm_config)
 
         if (
             parallel_config.distributed_executor_backend is not None
