+++ vllm_rbln/v1/worker/optimum_worker.py
@@ -38,7 +38,7 @@ from vllm.v1.worker.worker_base import WorkerBase
 
 import vllm_rbln.rbln_envs as envs
 from vllm_rbln.logger import init_logger
-from vllm_rbln.utils.optimum.converter import update_num_blocks
+from vllm_rbln.utils.optimum.converter.optimum_to_vllm import update_num_blocks
 from vllm_rbln.utils.optimum.params import RBLNParams
 from vllm_rbln.v1.worker.optimum_model_runner import RBLNOptimumModelRunner
 from vllm_rbln.v1.worker.utils import set_omp_num_threads
